requirements.txt
astroid==2.11.6
async-timeout==4.0.2
attrs==21.4.0
cachelib==0.9.0
click==8.1.3
coverage==6.4.1
coverage-badge==1.1.0
Deprecated==1.2.13
dill==0.3.5.1
Flask==2.1.2
Flask-Session==0.4.0
gunicorn==20.1.0
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary
py==1.11.0
pylint==2.14.4
pyparsing==3.0.9
pytest==7.1.2
redis==4.3.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
typing_extensions==4.3.0
Werkzeug==2.1.2
wrapt==1.14.1
